+++ internal/storage/migrations.go
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS Orders
     id SERIAL PRIMARY KEY,
     order_id varchar(50),
     status varchar(15),
-	amount INTEGER,
+	amount REAL,
 	uploaded_at timestamp with time zone,
 	processed_at timestamp with time zone,
 	user_id INTEGER REFERENCES Users(id)
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS Withdrawn
 (
     id SERIAL PRIMARY KEY,
     order_id varchar(50),
-    amount INTEGER,
+    amount REAL,
 	uploaded_at timestamp with time zone,
 	user_id INTEGER REFERENCES Users(id)
 );
@@ -31,8 +31,8 @@ CREATE TABLE IF NOT EXISTS Withdrawn
 CREATE TABLE IF NOT EXISTS Balance
 (
 	id      SERIAL PRIMARY KEY,
-	balance INTEGER,
-	withdrawn INTEGER,
+	balance REAL,
+	withdrawn REAL,
 	user_id INTEGER REFERENCES Users(id)
 );
 `
